Refactor apps into separate files
